feat: Setup uv, mise, tools, fmt, lint, and ci
This is a full starting environment for python, ansible, testinfra,
tofu, and terragrunt projects.

This is the base system for my future development setup for automating
all of my machines, hosts, and remote boxes.

In addition, this will serve as a basis for expanding this to running
hardened debian images, running and managing kubernetes clusters, etc.

This is also a strict TDD setup, where pytest, molecule, etc will be
used to test everything.

In addition, I plan to integrate lima into this setup as a docker host
on osx, as a VM on machines I want to develop on, etc.